requirements.txt
discord.py>=2.4.0
aiosqlite
Pillow
python-dotenv